+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 * Added a new validation that checks in playbooks for the usage of `DeleteContext` in **validate** command.
-
+* Fixed an issue in the **upload** command where it would try to upload content entities with unsupported versions.
 
 # 1.2.7
 * Changed input parameters in **find-dependencies** command.